+++ src/core/prelude/structures/sync-promise/spec.js
@@ -69,6 +69,45 @@ describe('core/prelude/structures/sync-promise', () => {
 		expect(await i).toBe(1);
 	});
 
+	it('double promise resolution', async () => {
+		expect(
+			await new SyncPromise((resolve) => {
+				resolve(1);
+				resolve(2);
+			})
+		).toBe(1);
+
+		expect(
+			await new SyncPromise((resolve) => {
+				resolve(new Promise((r) => setTimeout(() => r(1)), 100));
+				resolve(2);
+			})
+		).toBe(1);
+	});
+
+	it('double promise rejection', async () => {
+		try {
+			await new SyncPromise((resolve, reject) => {
+				reject(1);
+				reject(2);
+			});
+
+		} catch (err) {
+			expect(err).toBe(1);
+		}
+
+		try {
+			await new SyncPromise((resolve, reject) => {
+				reject(new Promise((r) => setTimeout(() => r(1)), 100));
+				reject(2);
+			});
+
+		} catch (err) {
+			expect(err).toBeInstanceOf(Promise);
+			expect(await err).toBe(1);
+		}
+	});
+
 	it('resolved `then` after `catch`', () => {
 		let
 			i = 1;